fix(mongo) ：解决mongo自定义ID与其他自动填充审计字段冲突问题
1. 暂时使用hutool雪花id填充自定义mongoId
2. 解决@CreateDate及@CreateBy与ID不兼容问题

+++ src/main/java/com/example/boot3scaffold/model/entities/mongo/MongoTest.java
@@ -1,5 +1,7 @@
 package com.example.boot3scaffold.model.entities.mongo;
 
+import java.math.BigInteger;
+
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -13,7 +15,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = false)
 @Document("mongo_test")
-public class MongoTest extends BaseMongoModel<Long> {
+public class MongoTest extends BaseMongoModel<BigInteger> {
 
     @Indexed
     private Long uid;
